Support more filetypes, more flexibility for files used
With these changes, the user can more easily select what files are used and which aren't. Due to these changes, the user is free to add their own ASCII or audio files (not limited to only .wav, but also for instance .mp3), or remove certain ASCII or audio files, to their own liking, without having to bother about file-naming conventions or file types. 

KNOWN ISSUE: files with a file name containing a space are not supported. This is due to the array splitting up the file name into different items.